+++ migration/index.spec.ts
@@ -22,7 +22,27 @@ describe("Migration", async () => {
   });
 
   tester.run(ruleName, theRule, {
-    valid: [`const routing = {};`],
-    invalid: [],
+    valid: [`new Integration({ typescript, config, routing });`],
+    invalid: [
+      {
+        name: "should import typescript and add it as a property to constructor argument",
+        code:
+          `import { Integration } from "express-zod-api";\n` +
+          `new Integration({ config, routing });`,
+        output:
+          `import typescript from "typescript";\n` +
+          `import { Integration } from "express-zod-api";\n` +
+          `new Integration({ typescript, config, routing });`,
+        errors: [
+          {
+            messageId: "add",
+            data: {
+              subject: "typescript property",
+              to: "constructor argument",
+            },
+          },
+        ],
+      },
+    ],
   });
 });